requirements.txt
annotated-types==0.7.0
anyio==4.7.0
fastapi==0.115.6
grpcio==1.67.1
idna==3.10
milvus-lite==2.4.10
numpy==2.2.0
pandas==2.2.3
protobuf==5.29.1
pydantic==2.10.3
pydantic_core==2.27.1
pymilvus==2.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
setuptools==75.6.0
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
transformers
sentence_transformers
sentencepiece==0.2.0
blobfile==3.0.0
lxml==5.3.0
pycryptodomex==3.21.0
tiktoken==0.8.0
datasets==3.2.0
uvicorn
accelerate==0.26.0
faiss_cpu==1.9.0
sentence-transformers==2.3.0
